requirements.txt
asgiref==3.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-appconf==1.1.0
django-cleanup==9.0.0
django-imagekit==5.0.0
idna==3.10
pilkit==3.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
